Feat : 선택된 그룹의 채널리스트 출력
