+++ client/src/components/SideBar/Channels/style.tsx
@@ -30,28 +30,6 @@ const ChannelType = styled.div`
   }
 `;
 
-const ChannelList = styled.ul`
-  & li {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    padding: 7px 10px;
-    margin: 2px 5px;
-    border-radius: 5px;
-    color: ${colors.Gray1};
-    &:hover {
-      background-color: ${colors.Gray3};
-      cursor: pointer;
-      color: ${colors.Black};
-    }
-    & div {
-      display: flex;
-      align-items: center;
-      & img {
-        margin-right: 5px;
-      }
-    }
-  }
-`;
+const ChannelList = styled.ul``;
 
 export { ChannelWrapper, ChannelType, ChannelList };
